+++ src/main/kotlin/com/manimdsl/runtime/datastructures/array/ArrayExecutor.kt
@@ -3,7 +3,10 @@ package com.manimdsl.runtime.datastructures.array
 import com.manimdsl.frontend.*
 import com.manimdsl.linearrepresentation.*
 import com.manimdsl.runtime.*
-import com.manimdsl.runtime.datastructures.DSExecutor
+import com.manimdsl.runtime.datastructures.BoundaryShape
+import com.manimdsl.runtime.datastructures.DataStructureExecutor
+import com.manimdsl.runtime.datastructures.SquareBoundary
+import com.manimdsl.runtime.datastructures.WideBoundary
 import com.manimdsl.runtime.utility.getBoundaries
 import com.manimdsl.stylesheet.Stylesheet
 
@@ -16,7 +19,7 @@ class ArrayExecutor(
     override val dataStructureBoundaries: MutableMap<String, BoundaryShape>,
     override val variableNameGenerator: VariableNameGenerator,
     override val codeTextVariable: String
-) : DSExecutor {
+) : DataStructureExecutor {
 
     override fun executeConstructor(node: ConstructorNode, dsUID: String, assignLHS: AssignLHS): ExecValue {
         val is2DArray = node.arguments.size == 2
